Moved cm_utils to the CM_module
